Implement ShellBrowserTestMainParts and consolidate shell_test_switches.h
