+++ cobalt/shell/BUILD.gn
@@ -252,10 +252,7 @@ static_library("cobalt_shell_lib") {
   if (run_browser_tests) {
     defines += [ "RUN_BROWSER_TESTS=1" ]
 
-    public_deps += [
-      ":shell_test_switches_lib",
-      "//content/test:test_support",
-    ]
+    public_deps += [ "//content/test:test_support" ]
   }
 
   if (shell_use_toolkit_views) {
@@ -300,15 +297,6 @@ static_library("cobalt_shell_lib") {
       "common/device_authentication.h",
     ]
   }
-}
-
-source_set("shell_test_switches_lib") {
-  testonly = true
-  sources = [
-    "common/shell_test_switches.cc",
-    "common/shell_test_switches.h",
-  ]
-  deps = [ "//base:base" ]
 }
 
 # This component provides a ContentMainDelegate for Content Shell and derived
