+++ cobalt/testing/browser_tests/browser/shell_browser_test_main_parts.cc
@@ -1,0 +1,41 @@
+// Copyright 2025 The Cobalt Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law-or-agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "cobalt/testing/browser_tests/browser/shell_browser_test_main_parts.h"
+
+#include "build/build_config.h"
+#include "cobalt/testing/browser_tests/common/shell_test_switches.h"
+
+#if BUILDFLAG(IS_LINUX)
+#include "ui/linux/linux_ui.h"          // nogncheck
+#include "ui/linux/linux_ui_factory.h"  // nogncheck
+#endif
+
+namespace content {
+
+ShellBrowserTestMainParts::ShellBrowserTestMainParts() = default;
+
+ShellBrowserTestMainParts::~ShellBrowserTestMainParts() = default;
+
+void ShellBrowserTestMainParts::ToolkitInitialized() {
+  if (switches::IsRunWebTestsSwitchPresent()) {
+    return;
+  }
+
+#if BUILDFLAG(IS_LINUX)
+  ui::LinuxUi::SetInstance(ui::GetDefaultLinuxUi());
+#endif
+}
+
+}  // namespace content
