+++ cobalt/testing/browser_tests/browser/test_shell.cc
@@ -46,7 +46,7 @@ TestShell* TestShell::CreateShell(std::unique_ptr<WebContents> web_contents,
   // Note: Do not make RenderFrameHost or RenderViewHost specific state changes
   // here, because they will be forgotten after a cross-process navigation. Use
   // RenderFrameCreated or RenderViewCreated instead.
-  if (test_switches::IsRunWebTestsSwitchPresent()) {
+  if (switches::IsRunWebTestsSwitchPresent()) {
     raw_web_contents->GetMutableRendererPrefs()->use_custom_colors = false;
     raw_web_contents->SyncRendererPrefs();
   }
@@ -151,14 +151,14 @@ bool TestShell::DidAddMessageToConsole(
     const std::u16string& message,
     int32_t line_no,
     const std::u16string& source_id) {
-  return test_switches::IsRunWebTestsSwitchPresent();
+  return switches::IsRunWebTestsSwitchPresent();
 }
 
 PictureInPictureResult TestShell::EnterPictureInPicture(
     WebContents* web_contents) {
   // During tests, returning success to pretend the window was created and allow
   // tests to run accordingly.
-  if (test_switches::IsRunWebTestsSwitchPresent()) {
+  if (switches::IsRunWebTestsSwitchPresent()) {
     return PictureInPictureResult::kSuccess;
   }
   return PictureInPictureResult::kNotSupported;
@@ -168,7 +168,7 @@ void TestShell::SetContentsBounds(WebContents* source,
                                   const gfx::Rect& bounds) {
   DCHECK(source == web_contents());  // There's only one WebContents per Shell.
 
-  if (test_switches::IsRunWebTestsSwitchPresent()) {
+  if (switches::IsRunWebTestsSwitchPresent()) {
     // Note that chrome drops these requests on normal windows.
     // TODO(danakj): The position is dropped here but we use the size. Web tests
     // can't move the window in headless mode anyways, but maybe we should be
