+++ cobalt/testing/browser_tests/media_browsertest.cc
@@ -22,7 +22,7 @@
 #include "build/build_config.h"
 #include "cobalt/shell/browser/shell.h"
 #include "cobalt/shell/common/shell_switches.h"
-#include "cobalt/shell/common/shell_test_switches.h"  // nogncheck
+#include "cobalt/testing/browser_tests/common/shell_test_switches.h"
 #include "cobalt/testing/browser_tests/content_browser_test_utils.h"
 #include "content/public/browser/web_contents.h"
 #include "content/public/common/content_features.h"
