+++ cobalt/testing/browser_tests/renderer/shell_content_renderer_test_client.cc
@@ -19,9 +19,9 @@
 #include "base/functional/bind.h"
 #include "base/notreached.h"
 #include "base/task/single_thread_task_runner.h"
-#include "cobalt/shell/common/shell_test_switches.h"
 #include "cobalt/testing/browser_tests/common/main_frame_counter_test_impl.h"
 #include "cobalt/testing/browser_tests/common/power_monitor_test_impl.h"
+#include "cobalt/testing/browser_tests/common/shell_test_switches.h"
 #include "cobalt/testing/browser_tests/renderer/shell_render_frame_observer.h"
 #include "content/public/common/pseudonymization_util.h"
 #include "content/public/test/test_service.mojom.h"
