+++ cobalt/testing/browser_tests/renderer/shell_render_frame_observer.cc
@@ -15,8 +15,8 @@
 #include "cobalt/testing/browser_tests/renderer/shell_render_frame_observer.h"
 
 #include "base/command_line.h"
-#include "cobalt/shell/common/shell_test_switches.h"
 #include "cobalt/testing/browser_tests/common/render_frame_test_helper.mojom.h"
+#include "cobalt/testing/browser_tests/common/shell_test_switches.h"
 #include "cobalt/testing/browser_tests/renderer/render_frame_test_helper.h"
 #include "content/public/renderer/render_frame.h"
 #include "content/public/renderer/render_frame_observer.h"
